+++ R/dock_view.R
@@ -1,19 +1,31 @@
-#' Dock view widget
-#'
 #' Create a dock view widget
 #'
-#' @import htmlwidgets
+#' Creates an interactive dock view widget that enables flexible
+#' layout management with draggable, resizable, and dockable panels.
+#' This is a wrapper around the dockview.dev
+#' JavaScript library, providing a powerful interface for
+#' creating IDE-like layouts in Shiny applications or R Markdown documents.
+#'
 #' @param panels Widget configuration. Slot for \link{panel}.
-#' @param ... Other options. See \url{https://dockview.dev/docs/api/dockview/options}.
+#' @param ... Other options. See
+#' \url{https://dockview.dev/docs/api/dockview/options/}.
 #' @param theme Theme. One of
 #' \code{c("abyss", "dark", "light", "vs", "dracula", "replit")}.
+#' @param addTab Globally controls the add tab behavior. List with enable and callback.
+#' Enable is a boolean, default to FALSE and callback is a
+#' JavaScript function passed with \link[htmltwidgets]{JS}.
 #' @param width Widget width.
 #' @param height Widget height.
 #' @param elementId When used outside Shiny.
 #'
+#' @returns An HTML widget object.
+#'
 #' @export
 #' @examplesShinylive
-#' webr::install("dockViewR", repos = "https://rinterface.github.io/rinterface-wasm-cran/")
+#' webr::install(
+#'  "dockViewR",
+#'  repos = "https://rinterface.github.io/rinterface-wasm-cran/"
+#' )
 #' library(shiny)
 #' library(bslib)
 #' library(dockViewR)
@@ -87,6 +99,7 @@ dock_view <- function(
     "dracula",
     "replit"
   ),
+  addTab = list(enable = FALSE, callback = NULL),
   width = NULL,
   height = NULL,
   elementId = NULL
@@ -100,10 +113,71 @@ dock_view <- function(
   # check reference panels ids
   check_panel_refs(panels, ids)
 
+  if (addTab$enable) {
+    if (is.null(addTab$callback)) {
+      addTab$callback <- JS(
+        "(config) => {
+          const addPanel = (panel, api) => {
+            let internals = {
+              component: 'default',
+              params: {
+                content: panel.content,
+                addTab: panel.addTab
+              }
+            }
+
+            // Handle removable option. If no,
+            // use the default tab component without the close panel button.
+            if (!panel.removable) {
+              internals.tabComponent = 'default';
+            }
+            let props = { ...panel, ...internals }
+            return (api.addPanel(props))
+          }
+          const defaultPanel = (pnId) => {
+            return (`
+              <p>Exchange me by running:</p>
+              <p>removeUI(<br>
+                &nbsp;&nbsp;selector = \"#${pnId} > *\",<br>
+                &nbsp;&nbsp;multiple = TRUE<br>
+              )</p>
+              <p>shiny::insertUI(<br>
+                    &nbsp;&nbsp;selector = \"#${pnId}\",<br>
+                    &nbsp;&nbsp;where = \"beforeEnd\",<br>
+                    &nbsp;&nbsp;ui = \"your ui code here\"<br>
+              )</p>
+            `)
+          }
+          const dockId = config.containerApi.component
+            .gridview.element.closest('.dockview')
+            .attributes.id.textContent;
+          const pnId = `panel-${Date.now()}`
+          addPanel({
+            id: pnId,
+            title: 'Panel new',
+            inactive: false,
+            content: {
+              head: '',
+              singletons: [],
+              dependencies: [],
+              html: defaultPanel(dockId + '-' + pnId)
+            },
+            position: { referenceGroup: config.group.id, direction: 'within' }
+          }, config.containerApi);
+        }"
+      )
+    } else {
+      if (!is_js(addTab$callback)) {
+        stop("`callback` must be a JavaScript function.")
+      }
+    }
+  }
+
   # forward options using x
   x <- list(
     theme = theme,
     panels = panels,
+    addTab = addTab,
     ...
   )
 
@@ -136,24 +210,41 @@ dock_view <- function(
 
 #' Dock panel
 #'
-#' Create a dock panel
+#' Create a panel for use within a [dock_view()] widget.
+#' Panels are the main container components that can be docked, dragged,
+#' resized, and arranged within the dockview interface.
 #'
-#' @import htmlwidgets
 #' @param id Panel unique id.
 #' @param title Panel title.
 #' @param content Panel content. Can be a list of Shiny tags.
 #' @param active Is active?
+#' @param removable Is the panel removable? Default to FALSE.
 #' @param ... Other options passed to the API.
-#' See \url{https://dockview.dev/docs/api/dockview/panelApi}.
+#' See \url{https://dockview.dev/docs/api/dockview/panelApi/}.
 #' If you pass position, it must be a list with 2 fields:
 #' - referencePanel: reference panel id.
 #' - direction: one of `above`, `below`, `left`, `right` or `within`
-#' (`above`, `below`, `left`, `right` put the panel in a new group, while `within` puts the panel
-#' after its reference panel in the same group).
+#' (`above`, `below`, `left`, `right` put the panel in a new group,
+#' while `within` puts the panel after its reference panel in the same group).
 #' Position is relative to the reference panel target.
 #'
+#' @return A list representing a panel object to be consumed by
+#' \link{dock_view}:
+#' - id: unique panel id (string).
+#' - title: panel title (string).
+#' - content: panel content (`shiny.tag.list` or single `shiny.tag`).
+#' - active: whether the panel is active or not (boolean).
+#' - ...: extra parameters to pass to the API.
+#'
 #' @export
-panel <- function(id, title, content, active = TRUE, ...) {
+panel <- function(
+  id,
+  title,
+  content,
+  active = TRUE,
+  removable = FALSE,
+  ...
+) {
   # We can't check id uniqueness here because panel has no
   # idea of other existing panel ids at that point.
   id <- as.character(id)
@@ -162,6 +253,7 @@ panel <- function(id, title, content, active = TRUE, ...) {
     id = id,
     title = title,
     inactive = !active,
+    removable = removable,
     content = htmltools::renderTags(content)
   )
 
@@ -188,9 +280,15 @@ panel <- function(id, title, content, active = TRUE, ...) {
 #' @param expr An expression that generates a dock_view
 #' @param env The environment in which to evaluate \code{expr}.
 #' @param quoted Is \code{expr} a quoted expression (with \code{quote()})? This
-#'   is useful if you want to save an expression in a variable.
+#' is useful if you want to save an expression in a variable.
 #'
 #' @rdname dock_view-shiny
+#'
+#' @return \code{dockViewOutput} and `dock_view_output`
+#' return a Shiny output function that can be used in the UI definition.
+#' \code{renderDockView} and `render_dock_view` return a
+#' Shiny render function that can be used in the server definition to
+#' render a `dock_view` element.
 #'
 #' @export
 dockViewOutput <- function(outputId, width = "100%", height = "400px") {
@@ -216,7 +314,12 @@ renderDockView <- function(expr, env = parent.frame(), quoted = FALSE) {
   if (!quoted) {
     expr <- substitute(expr)
   } # force quoted
-  htmlwidgets::shinyRenderWidget(expr, dockViewOutput, env, quoted = TRUE)
+  htmlwidgets::shinyRenderWidget(
+    expr,
+    dockViewOutput,
+    env,
+    quoted = TRUE
+  )
 } #nocov end
 
 #' Alias to \link{renderDockView}