+++ R/pkg.R
@@ -3,4 +3,5 @@
 
 #' @importFrom shiny icon getDefaultReactiveDomain
 #' @importFrom stats setNames
+#' @import htmlwidgets
 NULL