+++ R/select_panel.R
@@ -1,18 +1,15 @@
-#' Remove Panel dynamically
-#' @param dock_id Dock unique id. When using modules the namespace is
-#' automatically added.
-#' @param id Id of the panel that ought to be removed.
+#' Select a panel dynamically
+#' @param dock_id Dock unique id. When using modules the namespace
+#' is automatically added.
+#' @param id Panel id.
 #' @param session shiny session object.
 #' See \url{https://dockview.dev/docs/api/dockview/panelApi/}.
 #' @export
 #' @return This function is called for its side effect.
 #' It sends a message to JavaScript through the current websocket connection,
 #' leveraging the shiny session object.
-remove_panel <- function(
-  dock_id,
-  id,
-  session = getDefaultReactiveDomain()
-) {
+select_panel <- function(dock_id, id, session = getDefaultReactiveDomain()) {
+  id <- as.character(id)
   panel_ids <- get_panels_ids(dock_id, session)
   if (!(id %in% panel_ids))
     stop(sprintf(
@@ -21,5 +18,9 @@ remove_panel <- function(
       id,
       paste(panel_ids, collapse = ", ")
     ))
-  session$sendCustomMessage(sprintf("%s_rm-panel", session$ns(dock_id)), id)
+
+  session$sendCustomMessage(
+    sprintf("%s_select-panel", session$ns(dock_id)),
+    list(id = id)
+  )
 }