+++ tests/testthat.R
@@ -1,4 +1,3 @@
 library(testthat)
 library(dockViewR)
 test_check("dockViewR")
-